client/agent/requirements.txt
flask
flask-cors
requests
python-socketio
websocket-client